ttfi1: numsym: terms with abstract numeric type work
there were a couple of issues with the previous formulation, most
egregious of which was the fact that the interpreters had been
specialized to a specific numeric type. the interpreters here are
generically defined.

in addition, for purposes of exposition and consistency, NumSym.tf1 has
been converted back to a closed term