dev: extensible deserialization
also update the examples to use the generic stack-safe fixpoint utility

TODO: currently it's somewhat of a pain to use the duplicating
interpreter. this is because, unlike haskell where we can guide the
compiler to lazily derive instances, scala seems to require the
programmer to spell out the implicits. figure out a way to ease this
burden.